initial commit - test baud rate
